\name{freealg-class}
\docType{class}
\alias{freealg-class}
\title{Class \dQuote{freealg}}
\description{The formal \proglang{S4} class for code{freealg} objects}
\section{Objects from the Class}{
  Formal class \dfn{freealg} is used for functions such as \code{drop()}
  which need a \proglang{S4} object.
}
\author{Robin K. S. Hankin}
\keyword{classes}
