\name{letters}
\alias{letters}
\alias{lettersymbols}
\alias{a}
\alias{b}
\alias{c}
\alias{d}
\alias{e}
\alias{f}
\alias{g}
\alias{h}
\alias{i}
\alias{j}
\alias{k}
\alias{l}
\alias{m}
\alias{n}
\alias{o}
\alias{p}
\alias{q}
\alias{r}
\alias{s}
\alias{t}
\alias{u}
\alias{v}
\alias{w}
\alias{x}
\alias{y}
\alias{z}
\alias{A}
\alias{B}
\alias{C}
\alias{D}
\alias{E}
\alias{F}
\alias{G}
\alias{H}
\alias{I}
\alias{J}
\alias{K}
\alias{L}
\alias{M}
\alias{N}
\alias{O}
\alias{P}
\alias{Q}
\alias{R}
\alias{S}
\alias{T}
\alias{U}
\alias{V}
\alias{W}
\alias{X}
\alias{Y}
\alias{Z}
\docType{data}
\title{Single-letter symbols}
\description{
  Variables \code{a}, \code{b},\ldots \code{z} and their inverses
  \code{A}-\code{Z} are given their \code{freealg} semantic meaning.
  }
\usage{data(lettersymbols)}
\details{

  Twenty-six variables a-z are defined as their \code{freealg} semantic
  equivalent:
  
  \preformatted{
    a <- as.mvp("a")
    ...
    z <- as.mvp("z")
  }

  Lowercase letters \code{c}, \code{q}, \code{t}, and uppercase letters
  \code{C}, \code{D}, \code{F}, \code{I}, \code{T}` might pose
  difficulties, which is why this (lazy-loaded) functionality is
  experimental.

  These objects can be generated by running script \code{inst/symb.Rmd},
  which includes some further discussion and technical documentation and
  creates file \code{lettersymbols.rda} which resides in the
  \code{data/} directory.

}
\author{Robin K. S. Hankin}
\examples{

(a+b)*(a-b)

}
\keyword{datasets}

