\name{letters}
\alias{letters}
\alias{lettersymbols}
\title{Single-letter symbols}
\description{
  Variables \code{a}, \code{b},\ldots \code{z} and their inverses
  \code{A}-\code{Z} are given their \code{freealg} semantic meaning.
  }

\details{

Sometimes it is convenient in an \R session to have all 26 letters a-z
and all 26 uppercase letters A-Z adopt their free algebra
interpretations.  This is accomplished by executing the following
commands:

  \preformatted{
    for(letter in c(letters,LETTERS)){
      eval(parse(text=paste(letter, ' <- as.freealg("',letter, '")')))
    }
  }
  
  Lowercase letters \code{c}, \code{q}, \code{t}, and uppercase letters
  \code{C}, \code{D}, \code{F}, \code{I}, \code{T} might pose
  difficulties.
  
  These objects can be generated by running script \code{inst/symb.Rmd},
  which includes some further discussion and technical documentation and
  creates file \code{lettersymbols.rda} which formerly resided in the
  \code{data/} directory.

}
\author{Robin K. S. Hankin}
\keyword{datasets}

